Move Edition into span crate
