+++ crates/hir-expand/src/declarative.rs
@@ -1,8 +1,8 @@
 //! Compiled declarative macro expanders (`macro_rules!`` and `macro`)
 use std::sync::OnceLock;
 
-use base_db::{CrateId, Edition, VersionReq};
-use span::{MacroCallId, Span};
+use base_db::{CrateId, VersionReq};
+use span::{Edition, MacroCallId, Span};
 use syntax::{ast, AstNode};
 use triomphe::Arc;
