+++ crates/project-model/Cargo.toml
@@ -25,6 +25,7 @@ itertools.workspace = true
 
 # local deps
 base-db.workspace = true
+span.workspace = true
 cfg.workspace = true
 paths = { workspace = true, features = ["serde1"] }
 stdx.workspace = true
